+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.11.0 (TBA)
 
+- Added WMTTemplates to WMTOperationUIData [(#162)](https://github.com/wultra/mtoken-sdk-ios/pull/162)
 - Extended PushParser to support parsing of inbox notifications [(#158)](https://github.com/wultra/mtoken-sdk-ios/pull/158)
 - Added statusReason to UserOperation [(#156)](https://github.com/wultra/mtoken-sdk-ios/pull/156)